Refactor: 공통된 메소드를 youtubeClassRoomUtils로 추출
